Updated translations
Fixing typos in both en_US and fr_FR as well as grammar in fr_FR. Tried
to make the cases (a vs. A) in fr_FR as similar to the cases in en_US.

This is not the full changes, I'm waiting to see if this change will be
approved to commit the full updated file.